api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.5.0",
    "fastapi>=0.122.0",
    "pydantic>=2.12.4",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "supabase>=2.24.0",
    "uvicorn[standard]>=0.38.0",
]
